+++ packages/mobile-visualization/CHANGELOG.md
@@ -8,6 +8,14 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 3.4.0-beta.14 (1/22/2026 PST)
+
+#### 🚀 Updates
+
+- Add chart Legend component. [[#302](https://github.com/coinbase/cds/pull/302)]
+- Add support for hideBeaconLabels in Scrubber. [[#302](https://github.com/coinbase/cds/pull/302)]
+- Add support for custom bar components. [[#302](https://github.com/coinbase/cds/pull/302)]
+
 ## 3.4.0-beta.13 (1/20/2026 PST)
 
 #### 🚀 Updates